2-7 Saturday ,  1.0.1 Version

+++ SugarNursing/NSString+dateFormatting.m
@@ -45,7 +45,7 @@
     NSDateComponents *comps = [gregorian components:unitFlags fromDate:birthDay  toDate:now  options:0];
     NSInteger year = [comps year];
     
-    return [NSString stringWithFormat:@"%ld",year];
+    return [NSString stringWithFormat:@"%ld",(long)year];
 }
